+++ R/tokenlist.R
@@ -1,18 +1,32 @@
-new_tokenlist <- function(x = list(), tokens = character()) {
+new_tokenlist <- function(x = list(), lemma = NULL, pos, tokens = character()) {
   vec_assert(x, list())
+  if (!(is.null(lemma) | is.list(lemma))) {
+    rlang::abort("`lemma` must be NULL or a list.")
+  }
+  if (!(is.null(pos) | is.list(pos))) {
+    rlang::abort("`pos` must be NULL or a list.")
+  }
   vec_assert(tokens, character())
-  new_vctr(x, tokens = tokens, class = "textrecipes_tokenlist")
+  new_vctr(x, lemma = lemma, pos = pos, tokens = tokens, 
+           class = "textrecipes_tokenlist")
 }
 
-tokenlist <- function(x = list()) {
+tokenlist <- function(x = list(), lemma = NULL, pos = NULL) {
   x <- vec_cast(x, list())
-  new_tokenlist(x, tokens = unique(unlist(x)))
+  if (!is.null(lemma)) {
+    lemma <- vec_cast(lemma, list())
+  }
+  if (!is.null(pos)) {
+    pos <- vec_cast(pos, list())
+  }
+  new_tokenlist(x, lemma = lemma, pos = pos, tokens = unique(unlist(x)))
 }
 
 is_tokenlist <- function(x) {
   inherits(x, "textrecipes_tokenlist")
 }
 
+#' @export
 format.textrecipes_tokenlist <- function(x, ...) {
   out <- formatC(lengths(vec_data(x)))
   paste0("[", out, " tokens]")
@@ -26,8 +40,19 @@ obj_print_footer.textrecipes_tokenlist <- function(x, ...) {
   cat("# Unique Tokens: ", format(length(attr(x, "tokens"))), "\n", sep = "")
 }
 
-vec_restore.textrecipes_tokenlist <- function(x, to, ..., i = NULL) {
-  new_tokenlist(x, unique(unlist(x)))
+#' @export
+`[.textrecipes_tokenlist`<- function(x, i) {
+  if (is.null(attr(x, "lemma"))) {
+    lemma <- NULL
+  } else {
+    lemma <- attr(x, "lemma")[i]
+  }
+  if (is.null(attr(x, "pos"))) {
+    pos <- NULL
+  } else {
+    pos <- attr(x, "pos")[i]
+  }
+  tokenlist(vec_data(x)[i], lemma = lemma, pos = pos)
 }
 
 # Takes a vector of character vectors and keeps (for keep = TRUE) the words
@@ -44,10 +69,28 @@ tokenlist_filter <- function(x, dict, keep = FALSE) {
   seq_x <- seq_along(x)
   i <- rep(seq_x, lengths(x))
   j <- match(unlist(x), dict)
+
+  keep_id <- !is.na(j)
+  split_id <- factor(i[keep_id], seq_x)
   
-  out <- split(dict[j[!is.na(j)]], factor(i[!is.na(j)], seq_x))
+  out <- split(dict[j[keep_id]], split_id)
   names(out) <- NULL
-  new_tokenlist(out, dict)
+  
+  if (!is.null(attr(x, "lemma"))) {
+    lemma <- split(unlist(attr(x, "lemma"))[keep_id], split_id)
+    names(lemma) <- NULL
+  } else {
+    lemma <- NULL
+  }
+  
+  if (!is.null(attr(x, "pos"))) {
+    pos <- split(unlist(attr(x, "pos"))[keep_id], split_id)
+    names(pos) <- NULL
+  } else {
+    pos <- NULL
+  }
+  
+  new_tokenlist(out, lemma = lemma, pos = pos, tokens = dict)
 }
 
 tokenlist_apply <- function(x, fun) {
@@ -73,4 +116,40 @@ tokenlist_to_dtm <- function(x, dict) {
   
   out@Dimnames[[2]] <- dict
   out
+}
+
+tokenlist_lemma <- function(x) {
+  tokenlist(attr(x, "lemma"))
+}
+
+tokenlist_pos_filter <- function(x, pos_tags) {
+  if (!is_tokenlist(x)) {
+    rlang::abort("Input must be a tokenlist.")
+  }
+  
+  if (is.null(attr(x, "pos"))) {
+    rlang::abort("pos attribute not avaliable.")
+  }
+  
+  seq_x <- seq_along(x)
+  i <- rep(seq_x, lengths(x))
+  j <- match(unlist(attr(x, "pos")), pos_tags)
+  
+  keep_id <- !is.na(j)
+  split_id <- factor(i[keep_id], seq_x)
+  
+  out <- split(unlist(vec_data(x))[keep_id], split_id)
+  names(out) <- NULL
+  
+  pos <- split(unlist(attr(x, "pos"))[keep_id], split_id)
+  names(pos) <- NULL
+  
+  if (!is.null(attr(x, "lemma"))) {
+    lemma <- split(unlist(attr(x, "lemma"))[keep_id], split_id)
+    names(lemma) <- NULL
+  } else {
+    lemma <- NULL
+  }
+  
+  tokenlist(out, lemma = lemma, pos = pos)
 }